+++ CHANGELOG.md
@@ -11,10 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upcoming changes...
 
+## [0.4.0] - 2023-07-07
+
+### Added
+- Added long snippet generation check limit (`snippetLimit`)
+- Added command line option: `--snippet-limit` to support it
+
 ## [0.2.0] - 2023-07-04
 
 ### Added
-
 - First pass at the following Classes
     - Fingerprinting ([Winnowing](src/main/java/com/scanoss/Winnowing.java))
     - Scanning ([Scanner](src/main/java/com/scanoss/Scanner.java))
@@ -22,4 +27,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - JSON Utils ([JsonUtils](src/main/java/com/scanoss/utils/JsonUtils.java))
 - CLI ([CommandLine](src/main/java/com/scanoss/cli/CommandLine.java))
 
-[0.0.1]: https://github.com/scanoss/scanoss.java/compare/v0.0.0...v0.2.0
+[0.2.0]: https://github.com/scanoss/scanoss.java/compare/v0.0.0...v0.2.0
+[0.4.0]: https://github.com/scanoss/scanoss.java/compare/v0.2.0...v0.4.0